+++ packages/core/loglayer/src/types/index.ts
@@ -91,4 +91,12 @@ export interface LogLayerConfig {
    * If set to true, will not include metadata data in the log message.
    */
   muteMetadata?: boolean;
+  /**
+   * If set to true, child loggers will link to their parent's context instead of
+   * creating a shallow copy. This means changes to the context in child loggers
+   * will affect the parent logger's context and vice versa.
+   *
+   * Default is false.
+   */
+  linkParentContext?: boolean;
 }